+++ tf-modules/cumulus-rds-tf/main.tf
@@ -13,6 +13,38 @@ provider "aws" {
     key_prefixes = ["gsfc-ngap"]
   }
 }
+
+locals {
+  # Determine if snapshot SGs should be used
+  use_snapshot_sg = (
+    var.snapshot_identifier != null && var.use_snapshot_security_group
+  )
+
+  # Determine which base SG to use:
+  # 1. If security_group_id is provided, use it.
+  # 2. If snapshot is being restored and use_snapshot_security_group = true, don't specify (null).
+  # 3. Otherwise, use the new SG created by this module.
+  rds_security_group_id = try(
+    var.input_security_group_id != null
+      ? var.input_security_group_id
+      : (
+          local.use_snapshot_sg
+          ? null
+          : aws_security_group.rds_cluster_access.id
+        ),
+    null
+  )
+
+  # Effective SG ID:
+  # If null, attempt to read the actual SG attached to the cluster (useful when using snapshot SGs)
+  effective_security_group_id = try(
+    local.rds_security_group_id != null
+      ? local.rds_security_group_id
+      : aws_rds_cluster.cumulus.vpc_security_group_ids[0],
+    null
+  )
+}
+
 resource "aws_db_subnet_group" "default" {
   name_prefix = var.aws_db_subnet_group_prefix
   subnet_ids  = var.subnets
@@ -91,8 +123,8 @@ resource "aws_rds_cluster" "cumulus" {
   engine                  = "aurora-postgresql"
   engine_version          = var.engine_version
   database_name           = "postgres"
-  master_username         = var.db_admin_username
-  master_password         = var.db_admin_password
+  master_username         = var.snapshot_identifier == null ? var.db_admin_username : null
+  master_password         = var.snapshot_identifier == null ? var.db_admin_password : null
   backup_retention_period = var.backup_retention_period
   preferred_backup_window = var.backup_window
   db_subnet_group_name    = aws_db_subnet_group.default.id
@@ -103,7 +135,9 @@ resource "aws_rds_cluster" "cumulus" {
     max_capacity = var.max_capacity
     min_capacity = var.min_capacity
   }
-  vpc_security_group_ids          = [aws_security_group.rds_cluster_access.id]
+
+  vpc_security_group_ids = local.rds_security_group_id == null ? null : [local.rds_security_group_id]
+
   deletion_protection             = var.deletion_protection
   enable_http_endpoint            = true
   tags                            = var.tags